# api-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the release candidate version of the \"Swiss Corporate API\"\
    \ specification.\n"
  version: "1.0.0.2-SNAPSHOT"
  title: "Swiss Corporate API"
  contact:
    name: "SIX Group"
    url: "https://www.six-group.com/"
    email: "swisscorpapi@six-group.com"
host: "tbd.six-group.com"
basePath: "/api/bankingservices/corporate/v1"
...
schemes:
- "https"
paths:
...
  /oauth/token:
    post:
      tags:
      - "oauth"
      summary: "Get the OAuth access and refresh token"
      description: "Returns the OAuth access and refresh token for the specified client\
        \ and target\n"
      operationId: "oauthToken"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "grant_type"
        in: "formData"
        description: "either authorization_code or refresh_token"
        required: true
        type: "string"
        enum:
        - "authorization_code"
        - "refresh_token"
      - name: "code"
        in: "formData"
        description: "authorization code if applicable"
        required: false
        type: "string"
      - name: "refresh_token"
        in: "formData"
        description: "refresh token if applicable"
        required: false
        type: "string"
      - name: "redirect_uri"
        in: "formData"
        description: "original redirect uri if applicable"
        required: false
        type: "string"
      - name: "client_id"
        in: "formData"
        description: "client id"
        required: true
        type: "string"
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      - name: "X-PSU-IP-Address"
        in: "header"
        description: "IP address of the user initiating the operation"
        required: true
        type: "string"
      - name: "X-PSU-User-Agent"
        in: "header"
        description: "User agent of the user initiating the operation"
        required: true
        type: "string"
      responses:
        200:
          description: "OAuth token response"
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
            Cache-Control:
              type: "string"
              description: "must be: no-store"
              default: "no-store"
            Pragma:
              type: "string"
              description: "must be: no-cache"
              default: "no-cache"
          schema:
            $ref: "#/definitions/oauthTokenResponse"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Oauth"
...
definitions:
...
  oauthTokenResponse:
    type: "object"
    required:
    - "access_token"
    - "token_type"
    properties:
      access_token:
        type: "string"
      token_type:
        type: "string"
        default: "bearer"
      expires_in:
        type: "integer"
        minimum: 1
      refresh_token:
        type: "string"
    title: "oAuth Token Response"
    example:
      access_token: "access_token"
      refresh_token: "refresh_token"
      token_type: "bearer"
      expires_in: 1
...
  commonErrorResponse:
    type: "object"
    properties:
      type:
        $ref: "#/definitions/commonErrorType"
      title:
        type: "string"
        example: "This is the general problem description"
      detail:
        type: "string"
        example: "Detailed problem description with respect to the current request,\
          \ e.g., invalid account number format"
      instance:
        type: "string"
        example: "path/to/corresponding/resource"
    title: "Common Error Response"
  commonErrorType:
    type: "string"
    title: "Common Error Type"
    description: "Error Types for commonErrorResponse. See Appendix for detailed information."
    example: "/problems/TECHNICAL_ERROR"
    enum:
    - "/problems/INVALID_PAYLOAD"
    - "/problems/MALFORMED_PAYLOAD"
    - "/problems/INVALID_TOKEN"
    - "/problems/EXPIRED_TOKEN"
    - "/problems/INSUFFICIENT_PRIVILEGES"
    - "/problems/NO_ACCESS_TO_RESOURCE"
    - "/problems/RESOURCE_DOES_NOT_EXIST"
    - "/problems/RESOURCE_NOT_READY"
    - "/problems/RESOURCE_TOO_LARGE"
    - "/problems/WRONG_METHOD"
    - "/problems/OPERATION_NOT_ALLOWED"
    - "/problems/TECHNICAL_ERROR"
...